+++ CHANGELOG.md
@@ -15,7 +15,7 @@ Most recent version is listed first.
 - ong/client: Add log id http header: https://github.com/komuw/ong/pull/166
 
 ## v0.0.22
-- Panic middleware should include correct stack trace: https://github.com/komuw/ong/pull/164
+- Panic/recoverer middleware should include correct stack trace: https://github.com/komuw/ong/pull/164
 - Log client address without port: https://github.com/komuw/ong/pull/165
 
 ## v0.0.21